rasa-bot/requirements.txt
python-dotenv
transformers
torch
langchain
langchain-community
pinecone-client
